introduce specialized vodml instance writers and readers plus tests
I also added some basic logic for comparing vodmli xml documents
ignoring the order of elements, which required updating xmlunit
to version 2, and modifying some tests doing xml diffs.

Fixed test1 and test2 vodmli documents in svn submodule

Fixed an issue in the vodml-instance.xsd file as it didn't allow for
'-' in vodmlref prefixes.

This was a big commit, but hopefully it allows the next ones to
be smaller as I cover more and more examples and parts of the
specification.